requirements.txt
aiohttp
multidict
pynacl>=1.2.0
requests>=2.18.0
rsa>=3.4
secp256k1>=0.13.2
